requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
drf-spectacular>=0.15.1,<0.16
psycopg2>=2.8.6,<2.9
Pillow>=8.2.0,<8.3.0
djangorestframework-simplejwt>=5.2.2,<5.3.0ip
factory-boy>=3.3.0,<3.4.0
pytest  
pytest-django

